refactor(spec): split spec.go into responsibility-scoped files
